Add ContactsPolicy

+++ app/Http/Controllers/ContactsController.php
@@ -54,6 +54,7 @@ class ContactsController extends Controller
      */
     public function show(Contact $contact)
     {
+        $this->authorize('view', $contact);
         return view('contacts.show', compact('contact'));
     }
 
@@ -65,6 +66,7 @@ class ContactsController extends Controller
      */
     public function edit(Contact $contact)
     {
+        $this->authorize('view', $contact);
         return view('contacts.edit', compact('contact'));
     }
 
@@ -77,6 +79,7 @@ class ContactsController extends Controller
      */
     public function update(Request $request, Contact $contact)
     {
+        $this->authorize('view', $contact);
         $data = $request->validate([
             'name' => "required",
             'phone_number' => 'required|regex:/^\s*(?:\+?(\d{1,3}))?([-. (]*(\d{3})[-. )]*)?((\d{3})[-. ]*(\d{2,4})(?:[-.x ]*(\d+))?)\s*/',
@@ -96,6 +99,7 @@ class ContactsController extends Controller
      */
     public function destroy(Contact $contact)
     {
+        $this->authorize('view', $contact);
         $contact->delete();
         return redirect()->route('home');
     }
